+++ components/zcash_protocol/src/constants/testnet.rs
@@ -78,3 +78,31 @@ pub const HRP_UNIFIED_FVK: &str = "uviewtest";
 ///
 /// [zip-0316]: https://zips.z.cash/zip-0316
 pub const HRP_UNIFIED_IVK: &str = "uivktest";
+
+/// The HRP for a Bech32m-encoded testnet shielded-only Revision 2 Unified Address.
+///
+/// Defined in [ZIP 316][zip-0316].
+///
+/// [zip-0316]: https://zips.z.cash/zip-0316
+pub const HRP_UNIFIED_ADDRESS_R2: &str = "zutest";
+
+/// The HRP for a Bech32m-encoded testnet transparent-including Revision 2 Unified Address.
+///
+/// Defined in [ZIP 316][zip-0316].
+///
+/// [zip-0316]: https://zips.z.cash/zip-0316
+pub const HRP_UNIFIED_ADDRESS_R2_TI: &str = "tutest";
+
+/// The HRP for a Bech32m-encoded testnet Revision 2 Unified FVK.
+///
+/// Defined in [ZIP 316][zip-0316].
+///
+/// [zip-0316]: https://zips.z.cash/zip-0316
+pub const HRP_UNIFIED_FVK_R2: &str = "uvftest";
+
+/// The HRP for a Bech32m-encoded testnet Revision 2 Unified IVK.
+///
+/// Defined in [ZIP 316][zip-0316].
+///
+/// [zip-0316]: https://zips.z.cash/zip-0316
+pub const HRP_UNIFIED_IVK_R2: &str = "uvitest";